feat: 좋아요 기능 구현 및 인기글 조회 기능 구현 - 좋아요 기능 구현 - 커뮤니티 메인 페이지 인기글 부분 수정 (좋아요, 조회수 순으로 desc) - 커뮤니티 글 상세페이지 좋아요 부분 수정

+++ src/main/java/com/likelion/catdogpia/domain/dto/community/ArticleListDto.java
@@ -26,12 +26,12 @@ public class ArticleListDto {
     private Member member; //작성자
     private Attach attach; // 이미지
     private int viewCnt; //조회수
-    private int likeCnt; //좋아요
+    private Long likeCnt; //좋아요
     private Long commentCnt; //댓글수
     private LocalDateTime createdAt; //작성일
 
     @Builder
-    public ArticleListDto(Long id, Long categoryId, String title, Member member, Attach attach, int viewCnt, int likeCnt, Long commentCnt, LocalDateTime createdAt) {
+    public ArticleListDto(Long id, Long categoryId, String title, Member member, Attach attach, int viewCnt, Long likeCnt, Long commentCnt, LocalDateTime createdAt) {
         this.id = id;
         this.categoryId = categoryId;
         this.title = title;
